+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Common and useful classes, methods, exceptions etc.
 
 1. [BaseCollection] Treat the `null` index as "no index" only while adding new element, iow. do not treat empty
 string as "no index" behaviour.
+2. [Miscellaneous] [Regex] Use simpler & stronger pattern to match name of file
 
 # 1.1.0